devtools/client/memory/test/chrome/test_Heap_05.html

<!DOCTYPE HTML>
<html>
<!--
Test that we show a message when the census results are empty.
-->
<head>
    <meta charset="utf-8">
    <title>Tree component test</title>
    <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
    <div id="container"></div>
    <pre id="test">
        <script src="head.js" type="application/javascript;version=1.8"></script>
        <script type="application/javascript;version=1.8">
         window.onload = Task.async(function* () {
           try {
             const container = document.getElementById("container");

             yield renderComponent(Heap(immutableUpdate(TEST_HEAP_PROPS, {
               snapshot: immutableUpdate(TEST_HEAP_PROPS.snapshot, {
                 census: immutableUpdate(TEST_HEAP_PROPS.snapshot.census, {
                   report: {
                     bytes: 1,
                     totalBytes: 1,
                     count: 1,
                     totalCount: 1,
                     id: 1,
                     parent: undefined,
                     children: [
                       {
                         name: Cu.getJSTestingFunctions().saveStack(),
                         bytes: 1,
                         totalBytes: 1,
                         count: 1,
                         totalCount: 1,
                         children: undefined,
                         id: 2,
                         parent: 1,
                       },
                       {
                         name: "noStack",
                         bytes: 1,
                         totalBytes: 1,
                         count: 1,
                         totalCount: 1,
                         children: undefined,
                         id: 3,
                         parent: 1,
                       }
                     ]
                   },
                   display: censusDisplays.allocationStack,
                 }),
               }),
             })), container);

             ok(!container.querySelector(".empty"),
                "When the report is not empty, we should not show the empty message");

             // Empty Census Report

             const emptyCensus = {
               report: {
                 bytes: 0,
                 totalBytes: 0,
                 count: 0,
                 totalCount: 0,
                 id: 1,
                 parent: undefined,
                 children: undefined,
               },
               parentMap: Object.create(null),
               display: censusDisplays.allocationStack,
               filter: null,
               expanded: new Immutable.Set(),
               focused: null,
               state: censusState.SAVED,
             };

             yield renderComponent(Heap(immutableUpdate(TEST_HEAP_PROPS, {
               snapshot: immutableUpdate(TEST_HEAP_PROPS.snapshot, {
                 census: immutableUpdate(TEST_HEAP_PROPS.snapshot.census, emptyCensus),
               }),
             })), container);

             ok(container.querySelector(".empty"),
                "When the report is empty in census view, we show the empty message");
             ok(container.textContent.indexOf(L10N.getStr("heapview.empty")) >= 0);

             // Empty Diffing Report

             yield renderComponent(Heap(immutableUpdate(TEST_HEAP_PROPS, {
               view: { state: viewState.DIFFING, },
               diffing: {
                 firstSnapshotId: 1,
                 secondSnapshotId: 2,
                 census: emptyCensus,
                 state: diffingState.TOOK_DIFF,
               },
               snapshot: null,
             })), container);

             ok(container.querySelector(".empty"),
                "When the report is empty in diffing view, the empty message is shown");
             ok(container.textContent.indexOf(L10N.getStr("heapview.no-difference")) >= 0);

             // Empty Filtered Census

             yield renderComponent(Heap(immutableUpdate(TEST_HEAP_PROPS, {
               snapshot: immutableUpdate(TEST_HEAP_PROPS.snapshot, {
                 census: immutableUpdate(TEST_HEAP_PROPS.snapshot.census, immutableUpdate(emptyCensus, {
                   filter: "zzzz"
                 })),
               }),
             })), container);

             ok(container.querySelector(".empty"),
                "When the report is empty in census view w/ filter, we show the empty message");
             ok(container.textContent.indexOf(L10N.getStr("heapview.none-match")) >= 0);

           } catch(e) {
             ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
           } finally {
             SimpleTest.finish();
           }
         });
        </script>
    </pre>
</body>
</html>
